requirements.txt
flask
flask-cors
scikit-learn
numpy
opencv-python
google-api-python-client
python-dotenv
vaderSentiment
